pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-to-markdown-pipeline"
version = "0.1.0"
description = "Convert PDF documents to structured Markdown with a multi-stage pipeline."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Mehmet İşcan" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup :: Markdown",
]
dependencies = [
    "docling>=2.86.0",
    "markitdown[pdf]>=0.1.5",
    "PyYAML>=6.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=6.0",
    "ruff>=0.11",
    "pyright>=1.1",
    "pre-commit>=4.0",
]

[project.scripts]
pdf-pipeline = "scripts.run_pipeline:main"

[project.urls]
Homepage = "https://github.com/PhiniteLab/pdf-to-markdown-pipeline"
Repository = "https://github.com/PhiniteLab/pdf-to-markdown-pipeline"
Issues = "https://github.com/PhiniteLab/pdf-to-markdown-pipeline/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]

[tool.setuptools.package-data]
scripts = ["py.typed"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "RUF001"]

[tool.ruff.lint.isort]
known-first-party = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=scripts --cov-report=term-missing --cov-fail-under=70"
